feat: add mws handlers & mocks setup
