+++ package.json
@@ -9,10 +9,11 @@
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
     "format:fix": "prettier --write ./src",
-    "prepare": "husky"
+    "prepare": "husky",
+    "test": "vitest run --coverage"
   },
   "lint-staged": {
-    "*.{ts,tsx}": "eslint src/ --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
+    "*.{ts,tsx}": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0 --no-warn-ignored"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.6",
@@ -22,8 +23,9 @@
     "react-router-dom": "^6.24.1"
   },
   "devDependencies": {
-    "@testing-library/dom": "^10.3.1",
+    "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "@types/node": "^20.14.9",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
@@ -31,6 +33,7 @@
     "@typescript-eslint/eslint-plugin": "^7.13.1",
     "@typescript-eslint/parser": "^7.13.1",
     "@vitejs/plugin-react": "^4.3.1",
+    "@vitest/coverage-v8": "^2.0.3",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jsx-a11y": "^6.9.0",
@@ -39,11 +42,13 @@
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
     "husky": "^9.0.11",
+    "jsdom": "^24.1.0",
     "lint-staged": "^15.2.7",
+    "msw": "^2.3.1",
     "prettier": "^3.3.2",
     "typescript": "^5.2.2",
     "vite": "^5.3.1",
     "vite-plugin-eslint2": "^4.4.0",
-    "vitest": "^1.6.0"
+    "vitest": "^2.0.3"
   }
 }
