+++ src/services/api-service.ts
@@ -1,8 +1,6 @@
 import type { Params } from 'react-router-dom';
 import { rickAndMortyApi } from './rickandmorty-api';
 import { store } from './store';
-import { Character } from '@/interfaces';
-import { mockCharacters } from '@assets/mock';
 
 type IdParams = {
   params: Params<'characterId'>;
@@ -18,14 +16,11 @@ interface LoaderData {
   error?: string;
   name: string;
   page: string;
-  results?: Character[];
 }
 
 interface LoaderCharacter {
   id: string;
 }
-
-const mock = mockCharacters.results?.slice(0, 19);
 
 const characterLoader = async ({ params }: IdParams) => {
   if (!params.characterId) {
@@ -47,24 +42,23 @@ const characterLoader = async ({ params }: IdParams) => {
   return { id: params.characterId } as LoaderCharacter;
 };
 
-const charactersLoader = () => {
-  // const url = new URL(request.url);
-  // const name = url.searchParams.get('name')?.toLowerCase().trim() || '';
-  // const page = url.searchParams.get('page')?.toLowerCase().trim() || '1';
+const charactersLoader = async ({ request }: { request: Request }) => {
+  const url = new URL(request.url);
+  const name = url.searchParams.get('name')?.toLowerCase().trim() || '';
+  const page = url.searchParams.get('page')?.toLowerCase().trim() || '1';
 
-  // const promise = store.dispatch(
-  //   rickAndMortyApi.endpoints.getCharacters.initiate({ page, name })
-  // );
+  const promise = store.dispatch(
+    rickAndMortyApi.endpoints.getCharacters.initiate({ page, name })
+  );
 
-  // const info = await promise.unwrap();
-  // promise.unsubscribe();
+  const info = await promise.unwrap();
+  promise.unsubscribe();
 
   return {
-    // info: info.info,
-    // error: info.error,
-    // name,
-    // page,
-    results: mock,
+    info: info.info,
+    name,
+    error: info.error,
+    page,
   } as LoaderData;
 };
