+++ src/tests/node.ts
@@ -1,0 +1,4 @@
+import { setupServer } from 'msw/node';
+
+import { handlers } from './handlers';
+export const server = setupServer(...handlers);
