+++ src/tests/setupTests.ts
@@ -1,0 +1,7 @@
+import * as matchers from '@testing-library/jest-dom/matchers';
+import { cleanup } from '@testing-library/react';
+import { expect, afterEach } from 'vitest';
+
+expect.extend(matchers);
+
+afterEach(() => cleanup());
